add regex validation to item controller

+++ src/main/java/lk/ijse/spring/controller/ItemController.java
@@ -1,10 +1,12 @@
 package lk.ijse.spring.controller;
 
+import lk.ijse.spring.customStatusCodes.SelectedItemAndCustomerErrorStatus;
 import lk.ijse.spring.dto.ItemStatus;
 import lk.ijse.spring.dto.impl.ItemDTO;
 import lk.ijse.spring.exception.DataPersistException;
 import lk.ijse.spring.exception.ItemNotFoundException;
 import lk.ijse.spring.service.ItemService;
+import lk.ijse.spring.util.RegexProcess;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -36,6 +38,9 @@ public class ItemController {
     @PutMapping(value = "/{itemId}")
     public ResponseEntity<Void> updateItem(@PathVariable ("itemId") String itemId, @RequestBody ItemDTO itemDTO) {
         try {
+            if (!RegexProcess.itemIdValidation(itemId) || itemDTO == null) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             itemService.updateItem(itemId, itemDTO);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }catch (ItemNotFoundException exception) {
@@ -50,6 +55,9 @@ public class ItemController {
     @DeleteMapping(value = "/{itemId}")
     public ResponseEntity<Void> deleteItem(@PathVariable ("itemId") String itemId) {
         try {
+            if (!RegexProcess.itemIdValidation(itemId)) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             itemService.deleteItem(itemId);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }catch (ItemNotFoundException exception) {
@@ -63,6 +71,9 @@ public class ItemController {
 
     @GetMapping(value = "/{itemId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ItemStatus getSelectedItem(@PathVariable("itemId") String itemId) {
+        if (!RegexProcess.itemIdValidation(itemId)) {
+            return new SelectedItemAndCustomerErrorStatus(1, "Invalid Item Id");
+        }
         return itemService.getSelectedItem(itemId);
     }
